write function for finding mislabelled items
